requirements.txt
requests>=2.28.0
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0